extract async generators from cache logic
