+++ logic/cacheLogic.js
@@ -4,18 +4,14 @@ const TipOrderLogic = require('./tiporderLogic');
 const CommentLogic = require('./commentLogic');
 const TipLogic = require('./tipLogic');
 const BlacklistLogic = require('./blacklistLogic');
+const AsyncTipGeneratorsLogic = require('./asyncTipGeneratorsLogic');
 const axios = require('axios');
 const cache = require('../utils/cache');
 const BigNumber = require('bignumber.js');
-var AsyncLock = require('async-lock');
-var lock = new AsyncLock();
 const {getTipTopics, topicsRegex} = require('../utils/tipTopicUtil');
 const Util = require('../utils/util');
 const { Profile } = require('../models');
 const Fuse = require('fuse.js');
-const lngDetector = new (require('languagedetect'));
-lngDetector.setLanguageType('iso2');
-const cld = require('cld');
 
 const searchOptions = {
   threshold: 0.3,
@@ -65,41 +61,9 @@ module.exports = class CacheLogic {
   static async getTipsAndVerifyLocalInfo() {
     const tips = await aeternity.getTips();
 
-    // not await on purpose, just trigger background preview fetch
-    lock.acquire("LinkPreviewLogic.fetchAllLinkPreviews", async () => {
-      const previews = await LinkPreviewLogic.fetchAllLinkPreviews();
-      const tipUrls = [...new Set(tips.map(tip => tip.url))];
-      const previewUrls = [...new Set(previews.map(preview => preview.requestUrl))];
-
-      const difference = tipUrls.filter(url => !previewUrls.includes(url));
-
-      await difference.asyncMap(async (url) => {
-        await LinkPreviewLogic.generatePreview(url).catch(console.error);
-      })
-    });
-    let result = [];
-
-    await lock.acquire("TipLogic.fetchAllLocalTips", async () => {
-      const languages = await TipLogic.fetchAllLocalTips();
-      const tipIds = [...new Set(tips.map(tip => tip.id))];
-      const languageIds = [...new Set(languages.map(preview => preview.id))];
-
-      const difference = tipIds.filter(url => !languageIds.includes(url));
-
-      result = await difference.asyncMap(async (id) => {
-        let tip = tips.find(tip => tip.id === id)
-        let title = tip.title.replace(/[!0-9#.,?)-:'“@\/\\]/g, '');
-        //const probability = lngDetector.detect(title, 1);
-        const probability2 = await cld.detect(title).catch(() => ({}));
-        //const lang1 = probability[0] ? probability[0][0] !== null ? probability[0][0] : null : null;
-        const lang2 = probability2.languages ? probability2.languages[0].code : null;
-        return {id, lang2, title}
-      })
-      await TipLogic.bulkCreate(result.map(({id, lang2}) => ({
-        id,
-        language: lang2
-      })));
-    });
+    // not await on purpose, just trigger background actions
+    AsyncTipGeneratorsLogic.triggerGeneratePreviews(tips);
+    AsyncTipGeneratorsLogic.triggerLanguageDetection(tips);
 
     return tips;
   }
